\name{summary.Matchby}
\alias{summary.Matchby}
\alias{print.summary.Matchby}
\title{Summarizing output from Matchby}
\description{
  \code{\link{summary}} method for class \code{\link{Matchby}}
}
\usage{
  \method{summary}{Matchby}(object, ... , digits=5)
}
\arguments{
  \item{object}{An object of class "\code{Matchby}", usually, a
    result of a call to \code{\link{Matchby}}.}
  \item{digits}{The number of significant digits that should be
    displayed.}
  \item{...}{Other options for the generic summary function.}
}
\author{
  Jasjeet S. Sekhon, UC Berkeley, \email{sekhon@berkeley.edu},
  \url{https://www.jsekhon.com}.  
}
\seealso{ Also see \code{\link{Matchby}},  \code{\link{Match}}, \code{\link{GenMatch}},
  \code{\link{MatchBalance}}, 
  \code{\link{balanceUV}}, \code{\link{qqstats}}, \code{\link{ks.boot}},
  \code{\link{GerberGreenImai}}, \code{\link{lalonde}}
}
\keyword{htest}

%  LocalWords:  UC url seealso GenMatch MatchBalance balanceUV htest
%  LocalWords:  GerberGreenImai lalonde Matchby qqstats
