\name{summary.balanceMV}
\alias{summary.balanceMV}
\title{Summarizing output from balanceMV}
\description{
  \code{\link{summary}} method for class \code{\link{balanceMV}}
}
\usage{
  \method{summary}{balanceMV}(object, ..., digits=5)
}
\arguments{
  \item{object}{An object of class "\code{balanceMV}", usually, a
    result of a call to \code{\link{balanceMV}}.}
  \item{digits}{The number of significant digits that should be
    displayed.}
  \item{...}{Other options for the generic summary function.}  
}
\author{
  Jasjeet S. Sekhon, UC Berkeley, \email{sekhon@berkeley.edu},
  \url{http://sekhon.berkeley.edu/}.  
}
\seealso{
  Also see \code{\link{balanceMV}}, \code{\link{balanceUV}},
  \code{\link{summary.balanceUV}}, \code{\link{Match}}, \code{\link{GenMatch}},
  \code{\link{MatchBalance}}, \code{\link{ks.boot}}, \code{\link{GerberGreenImai}},
  \code{\link{lalonde}}
}
\keyword{htest}

%  LocalWords:  balanceMV UC url seealso balanceUV GenMatch MatchBalance htest
%  LocalWords:  GerberGreenImai lalonde
